Add QuestionImageManager module in QuizEditor
The commit includes creating a new 'QuestionImageManager' module where the image uploading functionality is migrated. Codes concerning the image uploading functionality are removed from 'QuizQuestionManager' and instead a call to the new 'QuestionImageManager' is placed. It results in cleaner and separated responsibilities.
